+++ Sources/MapboxMaps/Offline/OfflineManager+MapboxMaps.swift
@@ -2,6 +2,14 @@ import Foundation
 @_implementationOnly import MapboxCommon_Private
 
 extension MapboxCoreMaps.OfflineManager {
+
+    /// Construct a new offline manager
+    ///
+    /// - Parameter resourceOptions: ResourceOptions the resource options to manage.
+    public convenience init(resourceOptions: ResourceOptions) {
+        self.init(resourceOptions: MapboxCoreMaps.ResourceOptions(resourceOptions))
+    }
+
     /// Loads a new style package or updates the existing one.
     ///
     /// - Parameters: