+++ Tests/MapboxMapsTests/MapView/Integration Tests/MapViewIntegrationTestCase.swift
@@ -14,13 +14,11 @@ internal class MapViewIntegrationTestCase: IntegrationTestCase {
     internal override func setUpWithError() throws {
         try super.setUpWithError()
 
+        try guardForMetalDevice()
+
         guard let window = window,
               let rootView = rootViewController?.view else {
             throw XCTSkip("No valid UIWindow or root view controller")
-        }
-
-        guard MTLCreateSystemDefaultDevice() != nil else {
-            throw XCTSkip("No valid Metal device (OS version or VM?)")
         }
 
         let resourceOptions = ResourceOptions(accessToken: accessToken)
@@ -63,10 +61,7 @@ internal class MapViewIntegrationTestCase: IntegrationTestCase {
         mapView = nil
         style = nil
 
-        rootViewController?.viewWillDisappear(false)
-        rootViewController?.viewDidDisappear(false)
-        rootViewController = nil
-        window = nil
+        try super.tearDownWithError()
     }
 }