+++ Tests/MapboxMapsTests/MapView/Unit Tests/MapInitOptionsTests.swift
@@ -4,19 +4,19 @@ import XCTest
 class MapInitOptionsTests: XCTestCase {
     override func setUp() {
         super.setUp()
-        CredentialsManager.default.accessToken = ""
+        ResourceOptionsManager.default.resourceOptions.accessToken = "" // invalid token
     }
 
     override func tearDown() {
         super.tearDown()
-        CredentialsManager.default.accessToken = ""
+        ResourceOptionsManager.destroyDefault()
     }
 
     func testDefaultOptionsAreOverridden() {
         var updatedMapInitOptions = MapInitOptions()
         XCTAssertNotEqual(updatedMapInitOptions.resourceOptions.accessToken, "pk.aaaaaa")
 
-        CredentialsManager.default.accessToken = "pk.aaaaaa"
+        ResourceOptionsManager.default.resourceOptions.accessToken = "pk.aaaaaa"
 
         updatedMapInitOptions = MapInitOptions()
         XCTAssertEqual(updatedMapInitOptions.resourceOptions.accessToken, "pk.aaaaaa")